+++ core/src/main/java/com/novoda/location/provider/finder/GingerbreadLastLocationFinder.java
@@ -52,7 +52,7 @@ public class GingerbreadLastLocationFinder implements LastLocationFinder {
 	}
 
 	@Override
-	public Location getLastBestLocation(int minDistance, long minTime) {
+	public Location getLastBestLocation(float minDistance, long minTime) {
 		Location bestResult = null;
 		float bestAccuracy = Float.MAX_VALUE;
 		long bestTime = Long.MIN_VALUE;
